pyproject.toml
[project]
name = "d2c-ai-employee"
version = "0.0.1"
description = "An AI employee for D2C founders — substrate that lowers the cost of asking cross-tool questions"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.5",
    "pyyaml>=6.0",
    "python-dotenv>=1.0",
    "httpx>=0.27",
    "click>=8.1",
    "mcp>=1.0",
]

[project.scripts]
d2c = "d2c.cli.main:cli"
d2c-mcp = "d2c.mcp.server:main"

[dependency-groups]
dev = [
    "pytest>=8.0",
    "ruff>=0.5",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["d2c"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "N"]
